pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "LF_linearsys"
version = "0.1.0"
description = "FISTA solver for massive sparse linear systems in light field microscopy"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "matplotlib",
    "scipy",
    "scikit-image",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
where = ["src"]
